requirements.txt
streamlit==1.54.0
openai==2.17.0
python-dotenv==1.2.1
